fix(post-749 cycle-4): consolidate inspection enums + drop dead defenses
Cycle-4 review of PR #755 (HEAD 72dd5e353) surfaced 1 P1 + 4 INFO. All
addressed here.

WR-1 cycle-4 (P1): cycle-3 fix introduced TWO copies of INSPECTION_TYPES
+ INSPECTION_STATUSES constants and two near-identical enum-narrowing
helpers (narrowInspectionEnums in inspection-keys.ts, toInspection in
inspection-mutation-options.ts). Direct CLAUDE.md rule #3 violation
("No duplicate types -- search src/types/ before creating any type").
Extracted both constants, both type aliases (InspectionType +
InspectionStatus), and the narrowInspectionEnums helper to
src/types/sections/inspections.ts -- the canonical home next to the
Inspection interface that already encodes the same literal unions.
Re-bound Inspection.inspection_type and Inspection.status to the new
exported InspectionType / InspectionStatus aliases. Both consumer files
now import the single helper; their local copies dropped (saves ~70 LOC
of duplicate code and removes the triple-source-of-truth drift hazard
for any future status addition).

IN-1 cycle-4: fetchRevenueTrends now returns Array<Record<string,
unknown>> directly (via jsonArrayOrEmpty). The defensive
`(Array.isArray(raw) ? raw : []) as Array<...>` re-wraps at
financial-keys.ts:147-150 and report-keys.ts:121-124 are unreachable
under the new static type. Removed both: the call site now reads `const
rows = await fetchRevenueTrends(months);`.

IN-2 cycle-4: replaced `data!` non-null assertions at use-profile.ts:60
and use-profile-mutations.ts:51 with explicit `if (!data) throw new
Error("Profile not found");` guards. PostgREST's .single() can return
`{ data: null, error: null }` on connection-pool edge cases (which is
why .maybeSingle() exists); the assertion was documenting trust rather
than enforcing it.

IN-3 cycle-4: rewrote misleading omitUndefined JSDoc. Old text said "the
cast through unknown is required" which evoked the as-unknown-as pattern
CLAUDE.md rule #8 forbids. The implementation actually does a single-
step `as StripUndefined<T>` cast from a fully-narrowed Record. New
comment makes the soundness argument explicit: the loop above filters
every undefined-valued entry, so the remaining keys structurally satisfy
StripUndefined<T> and the cast is a sound widening, not a bridge.

IN-4 cycle-4: tightened useExpensesByProperty test with positive-shape
assertions on the filter columns -- a regression that swaps
units.eq("property_id", ...) for any other column would still leave the
table-sequence assertions intact. Now pins:
  expect(mockEq).toHaveBeenCalledWith("property_id", "prop-1")
  expect(mockIn).toHaveBeenNthCalledWith(1, "unit_id", ["unit-1"])
  expect(mockIn).toHaveBeenNthCalledWith(2, "maintenance_request_id", ["mr-1"])

100% typecheck clean, 0 lint errors, 106,111 unit tests pass.

[hudsor01]
