+++ src/hooks/api/query-keys/inspection-mutation-options.ts
@@ -11,8 +11,10 @@ import type {
 	UpdateInspectionInput,
 	UpdateInspectionRoomInput,
 } from "#lib/validation/inspections";
-import type { Inspection } from "#types/sections/inspections";
-import type { Tables } from "#types/supabase";
+import {
+	type Inspection,
+	narrowInspectionEnums,
+} from "#types/sections/inspections";
 
 export interface RecordPhotoInput {
 	inspection_room_id: string;
@@ -22,40 +24,6 @@ export interface RecordPhotoInput {
 	file_size?: number;
 	mime_type: string;
 	caption?: string;
-}
-
-const INSPECTION_TYPES = ["move_in", "move_out"] as const;
-const INSPECTION_STATUSES = [
-	"pending",
-	"in_progress",
-	"completed",
-	"tenant_reviewing",
-	"finalized",
-] as const;
-
-type InspectionType = (typeof INSPECTION_TYPES)[number];
-type InspectionStatus = (typeof INSPECTION_STATUSES)[number];
-
-// Narrows the DB row's `inspection_type` / `status: string` to the literal
-// unions the Inspection interface promises. DB CHECK constraints guarantee
-// the values at insert/update time; we reject anything else explicitly
-// rather than silently coercing.
-function toInspection(row: Tables<"inspections">): Inspection {
-	if (!INSPECTION_TYPES.includes(row.inspection_type as InspectionType)) {
-		throw new Error(
-			`Unexpected inspection_type "${row.inspection_type}" on inspection ${row.id}`,
-		);
-	}
-	if (!INSPECTION_STATUSES.includes(row.status as InspectionStatus)) {
-		throw new Error(
-			`Unexpected status "${row.status}" on inspection ${row.id}`,
-		);
-	}
-	return {
-		...row,
-		inspection_type: row.inspection_type as InspectionType,
-		status: row.status as InspectionStatus,
-	};
 }
 
 export const inspectionMutations = {
@@ -73,7 +41,7 @@ export const inspectionMutations = {
 					.single();
 
 				if (error) handlePostgrestError(error, "inspections");
-				return toInspection(created);
+				return narrowInspectionEnums(created);
 			},
 		}),
 
@@ -89,7 +57,7 @@ export const inspectionMutations = {
 					.single();
 
 				if (error) handlePostgrestError(error, "inspections");
-				return toInspection(updated);
+				return narrowInspectionEnums(updated);
 			},
 		}),
 
@@ -123,7 +91,7 @@ export const inspectionMutations = {
 					.single();
 
 				if (error) handlePostgrestError(error, "inspections");
-				return toInspection(updated);
+				return narrowInspectionEnums(updated);
 			},
 		}),
 
@@ -139,7 +107,7 @@ export const inspectionMutations = {
 					.single();
 
 				if (error) handlePostgrestError(error, "inspections");
-				return toInspection(updated);
+				return narrowInspectionEnums(updated);
 			},
 		}),
 
@@ -161,7 +129,7 @@ export const inspectionMutations = {
 					.single();
 
 				if (error) handlePostgrestError(error, "inspections");
-				return toInspection(updated);
+				return narrowInspectionEnums(updated);
 			},
 		}),
