+++ src/hooks/api/use-profile-mutations.ts
@@ -48,7 +48,8 @@ const profileMutationFactories = {
 					.select(PROFILE_SELECT)
 					.single();
 				if (error) throw error;
-				return mapUserProfile(data!);
+				if (!data) throw new Error("Profile not found");
+				return mapUserProfile(data);
 			},
 		}),
